# Run fmt parallel