Same as #208

Before
---

```console
❯ hyperfine dist/fmt
Benchmark 1: dist/fmt
  Time (mean ± σ):     196.8 ms ±   4.3 ms    [User: 93.4 ms, System: 136.8 ms]
  Range (min … max):   191.2 ms … 206.0 ms    14 runs
```

After
---

```console
❯ hyperfine dist/fmt
Benchmark 1: dist/fmt
  Time (mean ± σ):      98.8 ms ±   4.4 ms    [User: 114.5 ms, System: 149.6 ms]
  Range (min … max):    87.9 ms … 106.9 ms    28 runs
```